+++ src/main/groovy/ru/vyarus/gradle/plugin/mkdocs/task/MkdocsInitTask.groovy
@@ -3,8 +3,9 @@ package ru.vyarus.gradle.plugin.mkdocs.task
 import groovy.transform.CompileStatic
 import org.gradle.api.DefaultTask
 import org.gradle.api.GradleException
+import org.gradle.api.provider.Property
+import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.TaskAction
-import ru.vyarus.gradle.plugin.mkdocs.MkdocsExtension
 import ru.vyarus.gradle.plugin.mkdocs.util.TemplateUtils
 
 /**
@@ -15,13 +16,18 @@ import ru.vyarus.gradle.plugin.mkdocs.util.TemplateUtils
  * @since 13.11.2017
  */
 @CompileStatic
-class MkdocsInitTask extends DefaultTask {
+abstract class MkdocsInitTask extends DefaultTask {
+
+    /**
+     * Documentation sources folder (mkdocs sources root folder).
+     */
+    @Input
+    abstract Property<String> getSourcesDir()
 
     @TaskAction
     void run() {
-        MkdocsExtension extension = project.extensions.findByType(MkdocsExtension)
-
-        File dir = project.file(extension.sourcesDir)
+        String sourcesPath = sourcesDir.get()
+        File dir = project.file(sourcesPath)
         if (dir.exists() && dir.listFiles().length > 0) {
             throw new GradleException("Can't init new mkdocs site because target directory is not empty: $dir")
         }
@@ -29,8 +35,8 @@ class MkdocsInitTask extends DefaultTask {
         TemplateUtils.copy(project, '/ru/vyarus/gradle/plugin/mkdocs/template/init/', dir, [
                 projectName: project.name,
                 projectDescription: project.description ?: '',
-                docDir: extension.sourcesDir,
+                docDir: sourcesPath,
         ])
-        logger.lifecycle("Mkdocs site initialized: $extension.sourcesDir")
+        logger.lifecycle("Mkdocs site initialized: $sourcesPath")
     }
 }